requirements.txt
locust
python-decouple